\name{calc_var_total}
\alias{calc_var_total}
\title{
Calculate the total variance
}
\description{
Takes an expression matrix and calculates the total variance including all groups.
}
\usage{
calc_var_total(exp_matrix)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{exp_matrix}{
An expression matrix with the probes on the rows and the samples on the columns.
}
}
\value{
Returns a numeric vector
}
